requirements.txt
pygame
potracer
pillow
